refactor(plans): share the roadmap direct-fetch contract via useRoadmapFetch
EpicSubIssues re-implemented the same getBaseUrl('roadmap') -> fetch -> throw-on-not-ok -> useQuery({ retry: false, staleTime }) shape that useEpicBoardItem already carried. Extract it into a useRoadmapFetch helper both hooks build on, so the 'query roadmap directly, fall back on failure' contract lives in one place.
